feat: redirect / based on auth state and expand middleware protection
- Redirect / to /search if authenticated, /login if not
- Move middleware.ts to src/ to match project structure
- Expand mathcer to cover /search, /notifications, /user routes
- Add JWT_SECRET to .env.example
