+++ frontend/.env.example
@@ -5,3 +5,6 @@
 PORT_FRONTEND=XXXX
 PORT_BACKEND=XXXX
 NEXT_PUBLIC_API_URL=http://localhost:5000/api
+
+# JWT
+JWT_SECRET="your-secret-key-change-in-production"
